Finish sentry changes & fix more issues with Adapter
