--- primitives/src/adapter.rs
@@ -1,87 +1,0 @@
-use crate::{Address, BigNum, Channel, ValidatorId};
-use async_trait::async_trait;
-use serde::{Deserialize, Serialize};
-use std::{convert::From, fmt};
-
-pub type AdapterResult<T, AE> = Result<T, Error<AE>>;
-
-#[deprecated = "use adapter::primitives::Error"]
-pub trait AdapterErrorKind: fmt::Debug + fmt::Display {}
-#[deprecated = "use adapter::primitives::Deposit"]
-pub type Deposit = crate::Deposit<BigNum>;
-
-#[derive(Debug)]
-#[deprecated = "use adapter::primitives::Deposit"]
-pub enum Error<AE: AdapterErrorKind> {
-    Authentication(String),
-    Authorization(String),
-    /// Adapter specific errors
-    // Since we don't know the size of the Adapter Error we use a Box to limit the size of this enum
-    Adapter(Box<AE>),
-    /// You need to `.unlock()` the wallet first
-    LockedWallet,
-}
-
-impl<AE: AdapterErrorKind> std::error::Error for Error<AE> {}
-
-impl<AE: AdapterErrorKind> From<AE> for Error<AE> {
-    fn from(adapter_err: AE) -> Self {
-        Self::Adapter(Box::new(adapter_err))
-    }
-}
-
-impl<AE: AdapterErrorKind> fmt::Display for Error<AE> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Error::Authentication(error) => write!(f, "Authentication: {}", error),
-            Error::Authorization(error) => write!(f, "Authorization: {}", error),
-            Error::Adapter(error) => write!(f, "Adapter: {}", *error),
-            Error::LockedWallet => write!(f, "You must `.unlock()` the wallet first"),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[deprecated = "use adapter::Session"]
-pub struct Session {
-    pub era: i64,
-    pub uid: Address,
-}
-
-#[async_trait]
-pub trait Adapter: Send + Sync + fmt::Debug + Clone {
-    type AdapterError: AdapterErrorKind + 'static;
-
-    /// Unlock adapter
-    fn unlock(&mut self) -> AdapterResult<(), Self::AdapterError>;
-
-    /// Get Adapter whoami
-    fn whoami(&self) -> ValidatorId;
-
-    /// Signs the provided state_root
-    fn sign(&self, state_root: &str) -> AdapterResult<String, Self::AdapterError>;
-
-    /// Verify, based on the signature & state_root, that the signer is the same
-    fn verify(
-        &self,
-        signer: ValidatorId,
-        state_root: &str,
-        signature: &str,
-    ) -> AdapterResult<bool, Self::AdapterError>;
-
-    /// Get user session from token
-    async fn session_from_token<'a>(
-        &'a self,
-        token: &'a str,
-    ) -> AdapterResult<Session, Self::AdapterError>;
-
-    /// Gets authentication for specific validator
-    fn get_auth(&self, validator_id: &ValidatorId) -> AdapterResult<String, Self::AdapterError>;
-
-    /// Calculates and returns the total spendable amount
-    async fn get_deposit(
-        &self,
-        channel: &Channel,
-        address: &Address,
-    ) -> AdapterResult<Deposit, Self::AdapterError>;
-}
