+++ sentry/Cargo.toml
@@ -18,7 +18,7 @@ futures = "^0.3"
 async-std = "^1"
 async-trait = "^0.1"
 # Primitives
-primitives = { path = "../primitives", features = ["postgres"] }
+primitives = { path = "../primitives", features = ["postgres", "test-util"] }
 adapter = { version = "0.2", path = "../adapter" }
 chrono = { version = "0.4", features = ["serde"] }
 dashmap = "4"
